feat: #OBS-I321 : Renamed file names
