+++ api-service/src/controllers/DatasetMetrics/DatasetMetricsController.ts
@@ -1,38 +1,80 @@
 import { Request, Response } from "express";
-import _ from "lodash";
-import { executeNativeQuery } from "../../connections/druidConnection";
+import * as _ from "lodash"
 import { ResponseHandler } from "../../helpers/ResponseHandler";
-import vaidationSchema from "./DatasetMetricsValidationSchema.json"
-import { schemaValidation } from "../../services/ValidationService";
+import dayjs from 'dayjs';
+import { handleConnectors, handleDataFreshness, handleDataLineage, handleDataObservability, handleDataQuality, handleDataVolume } from "./datasetMetricsHelper";
 import logger from "../../logger";
-import { obsrvError } from "../../types/ObsrvError";
-import axios from "axios";
+import { schemaValidation } from "../../services/ValidationService";
+import validationSchema from "./DatasetMetrics.json";
 import { config } from "../../configs/Config";
+import { datasetService } from "../../services/DatasetService";
 
-const getBaseUrl = (url: string) => {
-    if (_.startsWith(url, "/prom")) return config.query_api.prometheus.url + _.replace(url, "/prom", "")
-}
-
+const apiId = "api.dataset.metrics";
 const datasetMetrics = async (req: Request, res: Response) => {
-    const isValidSchema = schemaValidation(req.body, vaidationSchema);
+    const msgid = _.get(req, "body.params.msgid");
+    const requestBody = req.body;
+    const dataset_id = _.get(req, "body.request.dataset_id");
+
+    const { category, volume_by_days = 10 }: any = req.body.request;
+    const defaultThreshold = (typeof config?.data_observability?.default_freshness_threshold === 'number' ? config?.data_observability?.default_freshness_threshold : 5) * 60 * 1000; // 5 minutes in milliseconds
+    const dateFormat = 'YYYY-MM-DDTHH:mm:ss';
+    const startDate = '2000-01-01T00:00:00+05:30';
+    const endDate = dayjs().add(1, 'day').format(dateFormat);
+    const intervals = `${startDate}/${endDate}`;
+    const isValidSchema = schemaValidation(requestBody, validationSchema);
+    const results = [];
+    
     if (!isValidSchema?.isValid) {
-        logger.error({ message: isValidSchema?.message, code: "INVALID_QUERY" })
-        throw obsrvError("", "INVALID_QUERY", isValidSchema.message, "BAD_REQUEST", 400)
+        logger.error({ apiId, datasetId: dataset_id, msgid, requestBody, message: isValidSchema?.message, code: "DATA_OUT_INVALID_INPUT" })
+        return ResponseHandler.errorResponse({ message: isValidSchema?.message, statusCode: 400, errCode: "BAD_REQUEST", code: "DATA_OUT_INVALID_INPUT" }, req, res);
     }
-    const { query } = req.body || {};
-    const endpoint = query.url;
-    if (_.startsWith(endpoint, "/prom")) {
-        query.url = getBaseUrl(endpoint)
-        const { url, method, headers = {}, body = {}, params = {}, ...rest } = query;
-        const apiResponse = await axios.request({ url, method, headers, params, data: body, ...rest })
-        const data = _.get(apiResponse, "data");
-        return res.json(data);
+
+    const dataset = await datasetService.getDataset(dataset_id, ["id"], true)
+    if (!dataset) {
+        logger.error({ apiId, message: `Dataset with id ${dataset_id} not found in live table`, code: "DATASET_NOT_FOUND" })
+        return ResponseHandler.errorResponse({ message: `Dataset with id ${dataset_id} not found in live table`, code: "DATASET_NOT_FOUND", statusCode: 404, errCode: "NOT_FOUND" }, req, res);
     }
-    else {
-        const query = _.get(req, ["body", "query", "body", "query"]);
-        const response = await executeNativeQuery(query);
-        ResponseHandler.successResponse(req, res, { status: 200, data: _.get(response, "data") });
+
+    try {
+        if (!category || category.includes("data_freshness")) {
+            const dataFreshnessResult = await handleDataFreshness(dataset_id, intervals, defaultThreshold);
+            results.push(dataFreshnessResult);
+        }
+
+        if (!category || category.includes("data_observability")) {
+            const dataObservabilityResult = await handleDataObservability(dataset_id, intervals);
+            results.push(dataObservabilityResult);
+        }
+
+        if (!category || category.includes("data_volume")) {
+            const dataVolumeResult = await handleDataVolume(dataset_id, volume_by_days, dateFormat);
+            results.push(dataVolumeResult);
+        }
+
+        if (!category || category.includes("data_lineage")) {
+            const dataLineageResult = await handleDataLineage(dataset_id, intervals);
+            results.push(dataLineageResult);
+        }
+
+        if (!category || category.includes("connectors")) {
+            const connectorsResult = await handleConnectors(dataset_id, intervals);
+            results.push(connectorsResult);
+        }
+
+        if (!category || category.includes("data_quality")) {
+            const connectorsResult = await handleDataQuality(dataset_id, intervals);
+            results.push(connectorsResult);
+        }
+
+        logger.info({ apiId, msgid, requestBody, datasetId: dataset_id, message: "Metrics fetched successfully" })
+        return ResponseHandler.successResponse(req, res, { status: 200, data: results });
+
     }
+    catch (error: any) {
+        logger.error({ apiId, msgid, requestBody: req?.body, datasetId: dataset_id, message: "Error while fetching metrics", code: "FAILED_TO_FETCH_METRICS", error });
+        return ResponseHandler.errorResponse({ message: "Error while fetching metrics", statusCode: 500, errCode: "FAILED", code: "FAILED_TO_FETCH_METRICS" }, req, res);
+    }
+
 }
 
 export default datasetMetrics;
