+++ api-service/src/routes/Router.ts
@@ -29,10 +29,10 @@ import ConnectorsRead from "../controllers/ConnectorsRead/ConnectorsRead";
 import DatasetImport from "../controllers/DatasetImport/DatasetImport";
 import { OperationType, telemetryAuditStart } from "../services/telemetry";
 import telemetryActions from "../telemetry/telemetryActions";
-import datasetMetrics from "../controllers/DatasetMetrics/DatasetMetricsController";
 import checkRBAC from "../middlewares/RBAC_middleware";
 import connectorRegisterController from "../controllers/ConnectorRegister/ConnectorRegisterController";
-import tableMetrics from "../controllers/TableMetrics/Metrics";
+import dataMetrics from "../controllers/DataMetrics/DataMetricsController";
+import datasetMetrics from "../controllers/DatasetMetrics/DatasetMetricsController";
 
 export const router = express.Router();
 
@@ -63,5 +63,5 @@ router.post("/datasets/import", setDataToRequestObject("api.datasets.import"), o
 router.post("/connector/register", setDataToRequestObject("api.connector.register"), onRequest({ entity: Entity.Management }), connectorRegisterController);
 //Wrapper Service
 router.post("/obsrv/data/sql-query", setDataToRequestObject("api.obsrv.data.sql-query"), onRequest({ entity: Entity.Data_out }), checkRBAC.handler(), sqlQuery);
-router.post("/data/metrics", setDataToRequestObject("api.data.metrics"), onRequest({ entity: Entity.Data_out }), datasetMetrics)
-router.post("/table/metrics/:dataset_id",setDataToRequestObject("api.table.metrics"), tableMetrics);
+router.post("/data/metrics", setDataToRequestObject("api.data.metrics"), onRequest({ entity: Entity.Data_out }), dataMetrics)
+router.post("/dataset/metrics", setDataToRequestObject("api.dataset.metrics"), checkRBAC.handler(), datasetMetrics);
